chore: optimize auth page
